update 22-10-15
update 22-10-15

+++ Workspace 2/TABS2/src/main/java/tabs/service/bidding/BidDataServiceBean.java
@@ -68,9 +68,17 @@ public class BidDataServiceBean implements BidDataService {
 	@Override
 	public void cancelBid(Long biddingId, int student_id) {
 		User u = userRepo.getUserByStudentID(student_id);
-		BidData persistedBidData = bidDataRepo.getUserBidData(biddingId, student_id);		
-		userRepo.updateToken(u.getStudent_id(), u.getToken()+persistedBidData.getToken());
-		bidDataRepo.deleteBidData(student_id, biddingId);						
+		BidData persistedBidData = bidDataRepo.getUserBidData(biddingId, student_id);
+		if(persistedBidData==null){
+			
+		}else{
+			userRepo.updateToken(u.getStudent_id(), u.getToken()+persistedBidData.getToken());
+			bidDataRepo.deleteBidData(student_id, biddingId);
+		}							
+	}
+	
+	public Collection<BidData> getWinBidDataList(Long bidding_id){		
+		return bidDataRepo.getWinBidDataList(bidding_id);
 	}
 
 }
